Доработка сброса пароля версия 3.5
