+++ src/main/java/com/avito/service/impl/UserServiceImpl.java
@@ -1,7 +1,9 @@
 package com.avito.service.impl;
 
+import com.avito.models.PasswordResetToken;
 import com.avito.models.User;
 import com.avito.models.posting.Posting;
+import com.avito.repository.PasswordResetTokenRepository;
 import com.avito.repository.PostingRepository;
 import com.avito.repository.UserRepository;
 import com.avito.service.interfaces.UserService;
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.util.Calendar;
 import java.util.List;
 
 @Service
@@ -24,7 +27,7 @@ public class UserServiceImpl implements UserService {
     private static final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
 
     private final PasswordEncoder passwordEncoder;
-
+    private final PasswordResetTokenRepository passwordResetTokenRepository;
     private final UserRepository userRepository;
     private final PostingRepository postingRepository;
 
@@ -73,5 +76,34 @@ public class UserServiceImpl implements UserService {
         Posting posting = postingRepository.getOne(id_posting);
         user.getFavoritePostings().remove(posting);
         userRepository.save(user);
+    }
+    @Override
+    public void addPasswordToken(PasswordResetToken passwordResetToken) {
+        passwordResetTokenRepository.save(passwordResetToken);
+    }
+    @Override
+    public String validatePasswordResetToken(String token) {
+        PasswordResetToken passToken = passwordResetTokenRepository.findByToken(token);
+        return !isTokenFound(passToken) ? "invalidToken"
+                : isTokenExpired(passToken) ? "expired"
+                : null;
+    }
+    private boolean isTokenFound(PasswordResetToken passToken) {
+        return passToken != null;
+    }
+
+    private boolean isTokenExpired(PasswordResetToken passToken) {
+        final Calendar cal = Calendar.getInstance();
+        return passToken.getExpiryDate().before(cal.getTime());
+    }
+    @Override
+    public void createPasswordResetTokenForUser(User user, String token) {
+        PasswordResetToken myToken = new PasswordResetToken(token, user);
+        addPasswordToken(myToken);
+    }
+
+    @Override
+    public PasswordResetToken findByToken(String token) {
+        return passwordResetTokenRepository.findByToken(token);
     }
 }
